Split the common parts of the Creep and PowerCreep into a Trait, so they can be handled by the same code where appropriate (e.g. movement code).
Not sure a if we should make this a Trait, or simply create a separate file and make it an abstract class.
